ضصثضصث

+++ dataTypes/error_handeling/exercise/calculate.py
@@ -1,8 +1,8 @@
 from divide import divide
 
 try:
-    num1 = int(input("Enter your first number: "))
-    num2 = int(input("Enter your second number: "))
+    num1 = float(input("Enter your first number: "))
+    num2 = float(input("Enter your second number: "))
 except ValueError as e:
     print(f"Error: invalid input! {e}")
 else:
